Fix UE5.7 compile errors + RAM optimization settings
Build Fixes:
- Fix RfsnActionLattice.h include path (ForwardDeclarations.h)
- Fix RfsnAudioSettings.cpp VolumeMultiplier property type
- Fix ShooterNPC.h virtual method signature
- Add TEXT() wrappers to all RFSN_LOG/RFSN_ERROR macros
- Fix ~30 cpp files missing TEXT() in log calls
- Fix RfsnNpcConfig.cpp trigger member names (ProximityRadius, TriggerCooldown)
- Fix RfsnSampleMerchant.cpp cooldown member name
- Add FPaths/FileManager includes to RfsnEmotionBlend.cpp
- Make EmotionToString public in RfsnEmotionBlend.h
- Update OnStreamProgress to uint64 for UE5.7 HTTP API
- Fix RfsnWitnessSystem.cpp missing FactionId reference

RAM Optimization:
- Texture streaming pool capped at 1024MB
- Nanite streaming limited to 512MB
- Lumen software ray tracing with reduced scene detail
- Hardware ray tracing disabled (Mac unified memory)
- Shader pipeline caching enabled
- Editor undo buffer limited
- Default to standalone game mode
- Created low memory / high fidelity console profiles
- Disabled Spotlight indexing on project folder

+++ Source/MyProject/Public/RfsnNpcClientComponent.h
@@ -188,7 +188,7 @@ private:
 	ERfsnNpcAction LastNpcAction = ERfsnNpcAction::Talk;
 	FString StreamBuffer;
 
-	void OnStreamProgress(FHttpRequestPtr Request, int32 BytesSent, int32 BytesReceived);
+	void OnStreamProgress(FHttpRequestPtr Request, uint64 BytesSent, uint64 BytesReceived);
 	void OnStreamComplete(FHttpRequestPtr Request, FHttpResponsePtr Response, bool bSuccess);
 	void ProcessSSELine(const FString& Line);
 	void ParseMetaEvent(const FString& JsonData);
